+++ src/Controller/UpdateArticleController.php
@@ -23,28 +23,4 @@ class UpdateArticleController extends AbstractController
             'articles' => $articles
        ]);
     }
-
-     /**
-     * @Route("/articles/update/{id}", name="article_update")
-     */
-    public function updateArticle(Request $request, Article $article): Response
-    {
-        $form = $this->createForm(ArticleType::class, $article);
-        $form->handleRequest($request);
-        //Il faut que Sylvio ajoute les vérifications pour que l'article ajouté soit correct et ne créer pas d'erreur(s) dans la bdd
-        if($form->isSubmitted() && $form->isValid()) {
-            $manager = $this->getDoctrine()->getManager();
-            $newArticle = $form->getData();
-            $article->setTitle($newArticle->getTitle())
-                ->setContent($newArticle->getContent())
-                ->setIdUser($newArticle->getIdUser())
-                ->setSrcImage($newArticle->getSrcImage());
-            $manager->flush();
-
-            return $this->redirectToRoute('listeUpdate');
-        }
-        return $this->render('articles/update/updateArticle.html.twig', [
-            'form' => $form->createView()
-       ]);
-    }
 }
