--- templates/articles/delete/deleteArticle.html.twig
@@ -1,15 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block title %}Supprimer un article{% endblock %}
-
-{% block body %}
-<table>
-{% for article in articles %}
-        <tr>
-            <td><div class="titreArticle"><h2>{{article.title}}</h2></div></td>
-            <td><div class="date">Ecrit le {{ article.createdAt | date('d/m/Y') }} à {{ article.createdAt | date('H:i') }}</div></td>
-            <td><a href="{{ path('article_delete', {'id': article.id}) }}">Supprimer</a></td>
-        </tr>
-{% endfor %}
-</table>
-{% endblock %}
