+++ templates/articles/listeArticles.html.twig
@@ -5,7 +5,7 @@
 {% block body %}
     {% for article in articles %}
         <div class="titreArticle"><p>{{ article.title }}</p></div>
-        <div class="boutonModifier"><a href="{{ path('article_update', {'id': article.id}) }}">Modifier</a>
-        <div class="boutonSupprimer"><a href="{{ path('article_delete', {'id': article.id}) }}">Supprimer</a></div>
+        <div class="boutonModifier"><a href="{{ path('update', {'id': article.id}) }}">Modifier</a>
+        <div class="boutonSupprimer"><a href="{{ path('delete', {'id': article.id}) }}">Supprimer</a></div>
     {% endfor %}
 {% endblock %}
