--- templates/articles/update/updateArticle.html.twig
@@ -1,11 +1,0 @@
-{% extends 'base.html.twig' %}
-
-{% block stylesheets %}
-    <link rel="stylesheet" type="text/css" href="{{ asset('css/article/update/updateArticle.css') }}">
-{% endblock %}
-
-{% block body %}
-    <div id='formUpdateArticle'>
-        {{ form(form) }}
-    </div>
-{% endblock %}
