+++ templates/header/header.html.twig
@@ -9,8 +9,7 @@
 				<nav>
 					<a href="{{ path('articles') }}" class="hvr-underline"> Articles </a>
 					<a href="{{ path('insert') }}" class="hvr-underline"> Créer </a>
-					<a href="{{ path('listeUpdate') }}" class="hvr-underline"> Modifier </a>
-					<a href="{{ path('delete') }}" class="hvr-underline"> Supprimer </a>
+					<a href="{{ path('liste_articles') }}" class="hvr-underline"> Gerer </a>
 				</nav>
 			</div>
 		</header>
